Use smart case or ignore case with grep commands
Although ag uses smart case by default, rg does not, and the other
programs do not support it. As this is a "fuzzyy" search, matching case
by default is likely to be surprising, so explicitly enable smart case
for programs that support it, and ignore case for those that don't.

FZF.vim, LeaderF.vim and telescope.nvim all enable smart case by default