Fix cursor restore in MacVim
Trying to restore cursor by saving and resetting &guicursor doesn't work
in MacVim r180 (Vim 9.1.727), when &guicursor starts at n:blinkon0 and
is restored after closing the popups, the cursor is no longer visible.

Fix by clearing the Cursor highlight group and restoring it instead.
This is similar to what LeaderF does (adds and then deletes a match).

Note that capturing the before state now happens when creating the popup
windows, not when loading popup.vim, to allow for color scheme changes.

P.S. Also tested that this works with hi-link, and with gVim on Windows.