Fix fallback to git-grep, git-ls-files after chdir
As the command strings for files and grep are computed at load time, the
command string is effectively cached, never computed again at runtime.

If the FuzzyyFiles or FuzzyyGrep commands are used first within a git
repo and the derived command is a git command, git-grep or git-ls-files,
that command will be used again after the user the changes the working
directory to somewhere that is not within a git repo, and won't work.

The same issue applies in reverse, user starts off in a non-git repo
directory, then changes to git repo, fallback git commands not used.

The fix for FuzzyyFiles is nice and neat, but fix for FuzzyyGrep is a
lot messier as a number of things change when the command changes. I'm
sure this can be improved, but better to fix first and improve later.