Add g:fuzzyy_grep_exclude_{file,dir}
This works the same as g:fuzzyy_files_exclude_{file,dir}, and uses the
same defaults. This allows the excludes for finding files and searching
in files to differ, which seems worthwhile, and is consistent with the
idea that each fuzzyy script can have its own configuration settings. A
separate default g:fuzzyy_excludes_{file,dir} could of course be added.