Prefer rg over ag and fd for Fuzzyy{Grep,Files}
Riprep is faster than ag, and almost as fast as fd, and making it the
default for both FuzzyFiles and FuzzyyGrep is more predictable (finding
files and searching in files will by default apply to the same files).

This is also the default in Telescope.nvim, and VSCode only uses ripgrep

Possible TODO: g:fuzzyy_{files,grep}_prefer_rg option, default true?