Include hidden files using ag & rg with FuzzyGrep
Hidden files are already included when using FuzzyFiles, and when using
other supported grep programs as they include hidden files by default,
so this makes Fuzzy behave more consistently for all external programs.

This is also, IMHO, the least surprising thing to do, for a plugin that
supports multiple programs for finding files and searching in files. As
a developer working in a codebase, I expect to find all the files that
have not been explicitly excluded/ignored, and I expect the same files
to be included whether I'm searching in files or searching for files.

It should be possible to override this if you really want to, but that's
for a separate commit (there are multiple ways to support that feature).