Include hidden files with PowerShell Get-ChildItem
For consistency with other programs used to find files. This enables
searching for files with the hidden attribute on Windows. UNIX style
hidden files, aka dotfiles, were already included by Get-ChildItem.

On Windows rg and fd consider both UNIX style hidden files and files
with the hidden attribute to be hidden, and excluded by default, but
included when using the --hidden option, so this change just makes
PowerShell Get-ChildItem include the same hidden files as rg and fd.