Respect global gitignore even when not in git repo
Use --no-require-git option with rg and fd when respecting gitignore

In future it will be possible to override all of these options, so this
will just be a friendly default. I think this is the least surprising
behaviour when the option to respect gitignore is enabled. Note that VS
code also enables this by default (I'm not sure you can even disable it)