Add option to follow symlinks
Unlike the hidden option this is disabled by default as it is not
supported by all programs, particularly git-ls-files and git-grep.

I'm undecided on this, and initially had it enabled by default as it
seemed like the least surprising thing to do, but concluded the lack of
support in git-ls-files and git-grep could cause confusion for users who
haven't installed alternative dependencies for FuzzyFiles and FuzzyGrep.