Add FuzzyMruCwd, remove g:fuzzyy_mru_project_only
This replaces the g:fuzzyy_mru_project_only option with a separate
command to start the MRU search within the current working directory.

CTRL-K to toggle scope still works, with both FuzzyMru and FuzzyMruCwd

Note that this changes the mru.Start() function to accept a starting or
working directory, which allows other variations to be easily added.

For example, you could have a variation for searching within a git repo,
which would be handy when working in large monorepos where the current
working directory might be a sub-directory from the repo root, e.g.

command! -nargs=0 FuzzyMruRepo mru.Start(windows.FuzzyMru, trim(system('git rev-parse --show-toplevel')))

The above is an example only, I do no propose adding this to the plugin
as is (though a smarter variation that considered other version control
systems and potentially other root markers might not be a bad idea).

P.S. I also think the signature for all Start() functions should be
standardised and use an optional opts dict as the only argument, with
the default window layout information read from a global or hard-coded.