Don't replace special buffers when opening files
This prevents Fuzzyy opening files in windows that are not intended for
editing, like those from file explorer plugins or the quickfix window.

This is normally what you want, you probably don't want to reuse these
special windows when opening a file for editing, for example opening a
file for editing in the NERDTree or similar file explorer drawer.

However, there are also some exceptions, especially if you like your
vinegar, and open some special buffers in the current editable window,
so this commit also adds a new option to set some exceptions (you may
recognise this from the venerable old CtrlP plugin, it's very similar).

See also http://vimcasts.org/blog/2013/01/oil-and-vinegar-split-windows-and-project-drawer/

Note: the implementation is a little fugly as there is no consistent way
to "open a file for editing" in Fuzzyy, so there are multiple calls to a
new function to move to the next usable window before opening files.