Disable results menu popup scrollbar by default
Including a scrollbar seems very much like an exception for a fuzzy
selector results menu, so change the menu scrollbar to off by default.

This was on by default before but disabled for most selectors, except
for FuzzyCommands, FuzzyCmdHistory, FuzzyInBuffer. After this change
it is also disabled for FuzzyCommands, FuzzyCmdHistory, FuzzyInBuffer.