Allow window layout customisation for all commands
Add cmdhistory, colors, commands, help, and inbuffer keys to window
layout dict, with pre-existing hard-coded values for widths and xoffset.

Note that as per other keys in this dict, it does not allow complete
customisation of the window layout, only overriding of the defaults
(e.g. you cannot currently set a height for any of the selectors).