Update Start() functions to use a single opts arg
This makes it easier to deal with multiple options for a selector, like
passing both a working directory and search keywords to grep.Start()

I want to do this for a custom FuzzyyGrepRoot command, which will
require adding an optional cwd argument to grep.Start(). Adding more
positional arguments to support new requirements like this is fugly.