Add option to customise async step value
This is used when asynchronously building the set of results, and avoids
calling matchfuzzypos() with massive lists, which really matters on code
repos with hundreds of thousands of files, otherwise the fuzzy matching
would seem unresponsive while waiting for matchfuzzypos() to return.

The default of 10,000 seems reasonable for most developer workstations,
but this is always going to be a compromise. You may want to reduce this
on an oldish or lower spec machine and notice a lack of responsiveness.