Add support for navigation using mouse
* Single click moves the cursor line in the menu window
* Double click selects a line in the menu window
* Mouse wheel scrolls up/down in the preview window

Scroll not implemented in menu as can require fetching new data, so a
bit tricky, select not implemented in preview as it's selector specific,
and might make no sense at all for some selectors (e.g. highlights).

Note: to accommodate scrolling for FuzzyFiles, the max number of lines
fetched when reading a file has been increased from 70 to 1000. Some
limit probably makes sense here, but it has been removed entirely from
FuzzyMru on the basis that if you've read the file before, you can read
it again (the same logic applies to FuzzyBuffers). FuzzyGrep also has no
limit, otherwise is would not always be possible to highlight the match.