Add FuzzyGrepRoot and FuzzyFilesRoot commands
This adds variants of FuzzyGrep and FuzzyFiles commands to search within
the "project" root, initially by simply looking for known source control
root markers in the filesystem, but could be extended with other checks.

The default root patterns are intentionally conservative, not including
patterns specific to build and packaging tools, programming languages,
editors etc. as it's very difficult to set defaults including these
things that are both widely applicable and will stand the test of time.
Some code repositories also contain multiple packages, so even defining
what is a project is not always obvious, probably best left to users.