Revised solution to warnings interrupting users
Move the code to show the warning to after the popups have been created,
which avoids the "Press ENTER or type command to continue" prompt when
the width of the message is wider than the current window (by the time
the message is echoed the current window is the fuzzyy prompt window)

But retain then use of echow for warnings re deprecated commands. These
warnings should only be shown when running the deprecated command, not
included in the shared list of deprecations for configuration options,
but we still need to avoid "Press ENTER or type command to continue".

We could pass the relevant warning(s) to selector.Start() and only show
one message, but that seems an overly complex solution to this problem,
and the combination of echow and echo actually works quite nicely.