Skip preview for binary files
Previewing binary files as text can mess up the screen, characters can
get rendered outside of the preview window which are not later cleared
when the popups are closed. It also looks pretty awful.

This adds a simple check for a NUL byte in the first 128 bytes, which is
fast and should be enough to detect a binary file. This is similar to
how ripgrep and other programs typically detect a binary, though they
may read the entire file. The 128 bytes might seem arbitrary, but it
comes from Leaderf, added in September 2023 and unchanged since then, so
presumably works well enough. Telescope.nvim does things differently,
actually running the file program to detect a mime type, but I wanted to
avoid that, at least until trying the simpler NUL byte check.

Note: as this code reads the first 128 bytes in binary mode and does no
encoding conversion, it will generate false positives for file encodings
that allow NUL bytes, like UTF-16 and UTF-32. This is IMHO an acceptable
trade-off to keep the code simple, Leaderf also detects encodings that
allow NUL bytes as binary, Telescope.nvim renders them as UTF-8, so with
a load of NUL byte marker garbage, git even considers them to be binary.
Rendering them in the Fuzzyy preview window would also include NUL byte
marker garbage as the file is read raw, without any encoding conversion.

Also note that this does not currently apply to the grep preview because
all the supported grep programs are configured to skip binaries anyway.