Code tidy, remove unnecessary v: variable prefix
Optional for null, true, and false in vim9script, including them doesn't
help readability. Also replace v:none as documented default value for x
and y offsets in window layout, this is incorrect, v:none is an empty
string, and using any non-number value will result in an error (the
defaults are actually automatically calculated to center the windows).