Ignore switchbuf=useopen when opening in split
Avoid using :sbuffer as it respects 'switchbuf=useopen', if you choose
to open a selected file in a split, then that's what should happen.

Before this change if 'switchbuf' included 'useopen' and you chose to
open a file in a split Fuzzyy would jump to an existing window rather
than actually open a new split if the file was already open somewhere.