Tweak MoveToUsableWindow for FuzzyBuffers
Always consider a window to be usable if the current buffer in the
window matches either the filetype or buftype of the one being opened.

For example, allow a terminal window to be reused to open a terminal
buffer even when 'terminal' is not included in g:fuzzyy_reuse_windows.