Fix jump to tag for some patterns, force nomagic
The tagaddress in a tags file should behave like 'magic' was not set,
but the previous code here just ran the Ex command without setting
'nomagic'. Tinkering with a user's global 'magic' setting is probably
bad, even if we do our utmost to reset it, so just update the code to
jump to a tag to work like the preview and force nomagic in patterns.

See https://docs.ctags.io/en/latest/man/tags.5.html and :help 'magic'