Add modeline fallback to detect preview filetype
Check modelines when filetype not detected using filetypedetect autocmd

Fuzzyy disables modelines for the preview window, which is sensible, you
don't want a modeline setting options and messing up the preview window.
But, this does mean that syntax highlighting is not enabled for some
files in the preview, but is enabled when opening the file for editing.

To address this, parse modelines if 'modeline' is on or securemodelines
plugin has been loaded (this is an old, but widely referenced plugin).

Note the check for either an empty &filetype or &filetype == 'conf'.
This is because 'conf' is used as the FALLBACK filetype in filetype.vim,
see :h :setfiletype and $VIMRUNTIME/filetype.vim, search for FALLBACK.

Note: Telescope.nvim also parse modelines, LeaderF just sets 'modeline'

TODO: refactoring, probably time to extract a PreviewFile() function