Add option to set custom devicon glyph function
Allows other plugins or custom functions to be used for devicons, e.g.
for vim-nerdfont `let g:fuzzyy_devicons_glyph_func = 'nerdfont#find'`

Note: vim-nerdfont does not follow the typical pattern for vim plugins
to set a "g:plugin_loaded" variable, and only uses autoloaded functions,
so there is no easy way for fuzzyy to check if it has been loaded.

Technical note: Fuzzyy measures the byte width for a single glyph and
assumes all glyphs returned by the devicon function are the same width
(this is almost certainly true, they should all be in the same Unicode
private use area so should use the same number of bytes in utf8, but
if you run into some weirdness with a custom function, could be this)