Fix incorrect results after paste into grep prompt
When pasting a string into the grep prompt the results shown in the menu
can be incorrect because Fuzzyy starts a new job for every character
entered, and although it ensures only one job is running, the output
callback can still receive data from the old channel momentarily.

Fix by checking the process id of the current job and the job associated
with the channel match before appending the output to the results.