Rename UpdateFzfList() to UpdateList()
And rename internal variable that this sets to raw_list, the use of fzf
as in variable and function names seems unnecessary, and confusing.

I'm doing this while pondering how to update the count of matches in the
menu title for FuzzyFiles, one option is exporting a getter function,
and I don't think including fzf in the function name or the internal
variable used as the return value is helpful.