Refactoring, do not export selector.windows
This is only used by FuzzyHighlights. While it's not necessarily a bad
idea, it is not the pattern followed elsewhere to get a reference to the
preview window, so replacing with that pattern for consistency. Longer
term it might just be better to always pass around a reference to the
windows dict rather than the menu window id as a callback argument.

Also renamed some vars to be consistent with naming of same elsewhere