Remove use of static counts as menu window title
These counts may be useful, but also a little confusing now that both
FuzzyFiles and FuzzyGrep show a dynamic count of matches in the title.

Showing a static count in some selectors and a dynamic count in others
is odd, and might seem like a bug, so remove the static ones for now.

I will revisit and aim to re-introduce these as dynamic counts later.