Don't add duplicate devicon highlight patterns
Functions to find a glyph for a filename pattern usually return a
default value when there is no specific match, which Fuzzyy uses to
apply a default color, by adding  a "__default__" filename pattern to
the color table, which will not be found and return the default glyph.

This is all very brittle though, and fails miserably when a new filename
pattern is added to the color table which also returns the default glyph
but applies a different color, so update the code to avoid applying the
same highlight pattern more than once, and sort the keys in the table
before processing them such that the default is always processed first.