Add vim-nerdtree-syntax-highlight compatibility
Automatically use colors from vim-nerdtree-syntax-highlight if installed
See https://github.com/tiagofumo/vim-nerdtree-syntax-highlight

This adds a compatibility script triggered via autocmds after Vim has
started (required to detect vim-nerdtree-syntax-highlight). Both
VimEnter and BufReadPre autocmds are required to allow for Vim starting
without or with a file passed as a command line argument.

This updates the devicons color table before the highlight groups are
created. To achieve this the color table has been moved out of the
devicons script to the colors script so it can be imported separately.

The color table exposed via a function which returns reference for
compatibility with Vim 9.0, which doesn't like the same variable name
being used for an exported variable and one referenced internally.