Use colorscheme's Cursor hl group where possible
Due to limitations in how popupwin operates, attempting to link the
`fuzzyyCursor` hlgroup to the `Cursor` hlgroup is normally not possible
as `Cursor` is hidden by Fuzzyy while the popup windows are active.

This works around that by resolving the `Cursor` properties and copying
them to `fuzzyyCursor` when loading the popups before hiding the cursor.

This is only done when `fuzzyCursor` is linked to `Cursor`, which is now
the default, but you can still override that default in your vimrc.

As the `Cursor` hlgroup is only intended for use with a gui cursor, gui
color values are resolved to roughly equivalent cterm color numbers if
possible, and otherwise fall back to using "reverse" as the highlight.

Co-authored-by: Mark Woods <mwoods.online.ie@gmail.com>

Closes #84