Use colorscheme's Cursor hl group for cursor
Due to limitations in how popupwin operates, attempting to use the
Cursor highlight group would not be possible as it is necessary to hide
it during popups.

This works around that by copying Cursor's resolved properties to an
internal highlight group, `fuzzyyPopup_cursor`, that will be used only
if `fuzzyCursor` is not linked to `Cursor`.