Handle only bg or fg set when resolving cursor hl
The previous code to resolve the cursor highlight didn't allow for only
the background or foreground to be set, i.e. only one of guifg & guibg.

This was done intentionally to keep the code simple, on the assumption
that only setting either background or foreground was so rare that it
didn't warrant the additional complexity, but it seems that assumption
was misguided, while it is unusual, it's not rare enough to ignore.

See https://github.com/Donaldttt/fuzzyy/pull/84 for background

Thanks @casr for reporting this and researching in some detail :-)