Add user autocmds to aid plugin compatibility
When using vim-lsp, language servers are started automatically when the
filetype is changed, which happens within the Fuzzyy preview window.

Adding user autocmds that are triggered when Fuzzyy opens and closes
allows vim-lsp to be disabled temporarily while Fuzzyy is open, e.g.

    augroup LspFuzzyy
      autocmd!
      autocmd User FuzzyyOpened call lsp#disable()
      autocmd User FuzzyyClosed call lsp#enable()
    augroup END

It turns out that setting `let g:lsp_untitled_buffer_enabled = 0`
achieves the same thing, but that is not exactly obvious from the
variable name (only looking at the code is it clear what this does).

Having said that, adding these user autocmds seems a good thing to allow
for compatibility with other plugins in general, not just for vim-lsp.

Note: this change replaces the PopupClosed user autocmd with FuzzyyClosed