Add FuzzyyPrevious command, reload previous search
This adds a feature similar to those found in Telescope, LeaderF and
FZF.vim to allow the previous selector and search to be reloaded.

A new launcher script is used to start selectors, and cache previous
launches, allowing the most recent one which also has some saved input
to be resumed. The cache currently only stores one previous launch (the
last one that had some associated input and can therefore be resumed).

This could be extended later to store a history of launches and allow
resumption of specific selectors, but just using the previous one with
some input is a good start (i.e. scratches my itch adequately for now).

The launcher script could also be used to support Fuzzy extensions in
future, allowing extensions to register selectors, kinda like Telescope.

The cache of previous launches does not persist across Vim sessions.
This is by design as persisting across sessions would need to take the
working directory into account, adding complexity, and potentially also
raises privacy concerns, so would probably need to be an opt-in feature.

Some minor changes were required to popup.SetPrompt() to support this.
It needs to clear any previously inserted text to support commands that
allow the prompt to be set, e.g. FuzzyGrep, otherwise the resumed input
is repeated; and needs to invoke the input callback after changing the
prompt to work reliably with async searches, otherwise the source list
may not have been fully loaded when fuzzy matching on resumed input.