Use file/buffer name as preview window title
When previewing a file or buffer, use tail of file name or buffer name
as the preview window title, similar to telescope dynamic_preview_title

This is always on at the moment in supported selectors as it seems a
good thing, just as showing a count of matches as the menu window title
where relevant in supported selectors is always on, and a good thing.

An option to disable it could be added, but doesn't seem necessary yet.