DRY code by adding async option to selector.Start()
Set async: true in opts to selector.Start() to automatically get async
search for selectors that don't need background jobs to build the list.

FuzzyHelp and FuzzyTags have been updated to use this new option. Though
fully functioning, consider it a work in progress, subject to change.

I have not enabled it for other selectors yet, as that would change the
current behaviour of those selectors, but have tested it with FuzzyMru
to ensure it works as expected when devicons are enabled.