Remove generic "title" opt from PopupSelection()
Would have set the menu title, but menu title is now set to the count of
matches, so this option was unused. Possibly to be replaced with a new
"menu_title" option, which would set the initial menu title. Not adding
that option yet as menu window might be renamed to "results".