Simplify devicons handling for most selectors
Move more of the code for handling devicons into core files so most
selectors don't need to worry about the details, set `devicons: true`
in opts to selector.Start() and it just works; devicons are added to
the results in the menu, and then removed from selected items passed
to preview, select and close callbacks.

Some custom handling remains within the files and grep selectors as they
currently require custom code to parse results when updating the menu.
Will revisit another time and see if I can remove the need for this.