Fix and simplify code to cache grep preview buffer
This wasn't actually working, and was unnecessarily complicated as the
main benefit to remembering the last path and linenr is to avoid having
to re-read a file into a buffer, once loaded jumping within the buffer
is plenty fast enough not to bother over-complicating the code.

The fix also removes the dependency on including current and last cursor
item in the opts argument, which should allow further simplification.