Remove opts dict from callback functions
Previously callback functions would need to look inside the opts dict to
find the cursor item and determine whether a preview window was loaded.

This is no longer necessary as the relevant window id and result(s) are
always provided where appropriate (close_cb does not get any results).

If it becomes necessary to add some kind of opts dict to callbacks in
future it can be done in a backwards compatible fashion, supporting
callback functions both declared either with and without an opts arg.