Move counter to prompt window as virtual text
And include count of matches alongside total count, where possible (this
doesn't work for live grep, so it continues to show only a match count)

There was a bit of fiddling with allowed argument types here to allow
for a non number value when using live grep (show a loading indicator),
that explains why the count and total args aren't limited to numbers.
This is messy, and it should never be exposed as a public api like this.