Fix check for launcher cache in launcher.Save()
Though the built-in commands now launch selectors via launcher.Start(),
this is not a requirement and custom selectors may be opened directly.

When this happens the cache may have been created, but be empty, so
update the code to allow for this and avoid index out or range warnings,
e.g.

    fuzzyy: Vim:E684: List index out of range: 0 at function
    <SNR>154_MenuFilter[61]..<SNR>154_GeneralPopupCallback[4]..fuzzyy#utils#launcher#Save,
    line 8

Fixes #98

Note: currently non-built in selectors must actually be opened directly
as the launcher has no way to find them, this will be addressed later.