Tweak FuzzyFiles menu timer for FuzzyPrevious
Increase the initial delay to 100ms to allow FuzzyPrevious reliably
trigger a menu update when feeding some keys. When Fuzzyy is run in
"minimalist" mode, no preview or devicons, the keys end up being fed
before the timer triggers the menu update, and the function to update
the menu thinks the input hasn't changed and doesn't update anything.

Yes, this is very hacky, but it does the job for now.