Replace key_callbacks with "actions"
Actions can be used to associate behaviour with keymaps. The difference
with key_callbacks is that they are invoked with pre-defined arguments,
the menu wid, the selected results, and opts used when creating popups.

The arguments allow the functions to be shared as they no longer rely
on script or global scope vars to execute commands in the window, access
the selected results, or obtain knowledge of the initialisation context
(e.g. to check the working directory or whether devicons are enabled).

Note: as part of this change, default actions are enabled for all
selectors to open files in splits, tabs, and send results to quickfix.
This keeps the code nice and simple, just works for future selectors
(built in or extensions) that list files, and is easily overridden.

Though I was initially wary of how this might behave for selectors that
don't list files, it doesn't actually blow up, it just does what you ask
it to do, even if that makes no sense (e.g. open helptags in quickfix).
Telescope.nvim also works this way, so it presumably doesn't cause much
confusion, or not enough to warrant added code complexity to avoid it.