Replace remaining key callbacks with actions
Key callbacks still supported for now, but not used by builtin selectors

Note: this update also allows action functions to be defined with no
args, one arg, two args, or three args, and it will be called correctly.
This is experimental, don't rely on it if writing any custom selectors.