Fix occasional bug loading grep preview
So, there used to be some code in the grep preview callback that tried
to be really smart and avoid loading the same file into the preview
window repeatedly when moving through multiple matches for the same file
in the menu, but it didn't actually work, in that the logic was wrong,
probably because it got broken somewhere along the line and was hard to
test manually, and had no automated tests (which would also be hard).

I tried to simplify and fix it, and it seemed to work fine, but there
are edge cases with timers and async callbacks both updating things, and
frankly it's just not worth bothering with the complexity, reading a
file and loading the preview for every selected result is plenty fast.

Long story short, premature optimisation is the root of all evil!