Allow default actions to be disabled
Fuzzbox now adds some default actions for opening files in splits and
tabs, and sending the results to the quickfix list. This is subject to
change, it was only really done to keep the code simple. Though it might
be useful to extension developers in future, that's not why it was done.

This commit just allows those actions to be disabled, by setting the
action for the keymap to the null function, or any non Funcref value.

Co-authored-by: Mark Woods <mwoods.online.ie@gmail.com>

Closes #109