Further simplify actions, change result to string
Previously the result arg was a list, to allow for a future where
Fuzzbox supports selecting multiple items and then acting on them
(the expectation being that if multiple selected, send to quickfix).

However, this complicates the code, and makes it harder to write simple
extensions as you always have to think about dealing with a list, so
replace the result list with a single result, the current cursor line.

When selecting multiple items is eventually supported I will worry about
select callbacks that want to handle multiple items (or maybe not worry,
acting on multiple selected items could be a different keymap & action).